Create MaxTicketsProfit
There are n ticket windows in the railway station. The with window has a(j) tickets available. The price of a ticket is equal to the number of tickets remaining in that window at that time. What is the maximum amount of money the railway station can earn from selling the first m tickets?

For example if we got 3 ticket windows have tickets of 3, 3, 4 and we want to sell 4 tickets. then

n = 3, m = 5
A[3] = {3, 3, 4}
The maximum money earned is 4 + 3 + 3 + 3 + 2 = 15